Adjusted file descriptions, added can_app to better interface, and resolved motor bit shift error.
