+++ firmware/components/drd/CMakeLists.txt
@@ -52,17 +52,17 @@ file(GLOB_RECURSE COMMON_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/../../common/*/*.c
 
 target_sources(${CMAKE_PROJECT_NAME} PRIVATE
     # Add user sources here
-    ${CMAKE_CURRENT_SOURCE_DIR}/app/evan_app.c
     ${CMAKE_CURRENT_SOURCE_DIR}/app/drive_state.c
     ${CMAKE_CURRENT_SOURCE_DIR}/app/soc.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/app/can_app.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/app/cyclic_data_handler.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/app/lcd_app.c
     ${CMAKE_CURRENT_SOURCE_DIR}/driver/accel_driver.c
     ${CMAKE_CURRENT_SOURCE_DIR}/driver/evan_driver.c
     ${CMAKE_CURRENT_SOURCE_DIR}/driver/gpio_driver.c
     ${CMAKE_CURRENT_SOURCE_DIR}/driver/can_driver.c
-    ${CMAKE_CURRENT_SOURCE_DIR}/app/lcd_app.c
     ${CMAKE_CURRENT_SOURCE_DIR}/driver/lcd_driver.c
     ${CMAKE_CURRENT_SOURCE_DIR}/driver/font_verdana.c
-    ${CMAKE_CURRENT_SOURCE_DIR}/app/cyclic_data_handler.c
     ${CMAKE_CURRENT_SOURCE_DIR}/runtime/tasks.c
     ${CMAKE_CURRENT_SOURCE_DIR}/runtime/interrupts.c
     ${COMMON_SOURCES}
@@ -73,8 +73,8 @@ target_include_directories(${CMAKE_PROJECT_NAME} PRIVATE
     # Add user defined include paths
     ${CMAKE_CURRENT_SOURCE_DIR}/app
     ${CMAKE_CURRENT_SOURCE_DIR}/driver
-    ${CMAKE_CURRENT_SOURCE_DIR}/../../common/cyclic_data/
     ${CMAKE_CURRENT_SOURCE_DIR}/runtime
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../common/cyclic_data/
     ${CMAKE_CURRENT_SOURCE_DIR}/../../common/debug_io
     ${CMAKE_CURRENT_SOURCE_DIR}/../../common/CAN_comms
 )
