--- firmware/components/drd/app/evan_app.c
@@ -1,6 +1,0 @@
-#include "evan_app.h"
-
-static void EvanAppFunction( void )
-{
-    int *p; 
-}
