--- firmware/components/drd/app/evan_app.h
@@ -1,6 +1,0 @@
-#ifndef __EVAN_DRIVER_H__
-#define __EVAN_DRIVER_H__
-
-#define EVAN_DRIVER 100
-
-#endif //__EVAN_DRIVER_H__
