+++ firmware/components/drd/driver/can_driver.c
@@ -4,36 +4,11 @@
  *
  * This file contains the implementation of CAN bus communication functions for the DRD board.
  * It handles CAN message transmission, reception, and filter configuration for vehicle state and control.
- *
- * @author  UBC Solar
- * @date    Feb 4 2026
  */
 
 /* INCLUDES */
 #include "can_driver.h"
-#include "CAN_comms.h"
 #include "can.h"
-#include "drive_state.h" // for now MMR change later
-
-/* FUNCTION PROTOTYPES */
-/**
- * @brief Handles received CAN messages for vehicle state and dispatches to appropriate handlers.
- *
- * @param msg_id The CAN message ID.
- * @param data Pointer to the received CAN message data.
- */
-void VehicleStateCanRxHandler(uint32_t msg_id, uint8_t* data);
-/**
- * @brief Initializes CAN hardware filters for message acceptance.
- *
- * @param can_filter Pointer to CAN filter configuration structure.
- */
-void CanFilterInit(CAN_FilterTypeDef* can_filter);
-/**
- * @brief Callback for processing received CAN messages.
- * @param CAN_comms_Rx_msg Pointer to the received CAN message structure.
- */
-void CanCommsRxCallback(CAN_comms_Rx_msg_t* CAN_comms_Rx_msg);
 
 /**
  *  CAN Message Header for drive control
@@ -92,62 +67,3 @@ void CanFilterInit(CAN_FilterTypeDef* can_filter) {
     can_filter2.FilterActivation = ENABLE;
     HAL_CAN_ConfigFilter(&hcan, &can_filter2); // Register filter with hardware
 }
-
-void CanTasksInit(void)
-{
-    CAN_comms_config_t CAN_comms_config_drd = {0};
-    CAN_FilterTypeDef can_filter = {0};
-    CanFilterInit(&can_filter);
-
-    CAN_comms_config_drd.hcan = &hcan;
-    CAN_comms_config_drd.CAN_Filter = can_filter;
-    CAN_comms_config_drd.CAN_comms_Rx_callback = CanCommsRxCallback;
-
-    CAN_comms_init(&CAN_comms_config_drd);
-}
-
-void CanCommsRxCallback(CAN_comms_Rx_msg_t* CAN_comms_Rx_msg)
-{
-	uint32_t CAN_ID = 0;
-	/*
-	 *	handle parsing rx messages
-	 */
-	if (CAN_comms_Rx_msg == NULL)
-	{
-			return;
-	}
-
-	if(CAN_comms_Rx_msg->header.IDE == CAN_ID_EXT)
-	{
-		CAN_ID = CAN_comms_Rx_msg->header.ExtId; // Get CAN ID
-	}
-	else
-	{
-		CAN_ID = CAN_comms_Rx_msg->header.StdId; // Get CAN ID
-	}
-
-    VehicleStateCanRxHandler(CAN_ID, CAN_comms_Rx_msg->data);
-}
-
-/* CAN RX */
-void VehicleStateCanRxHandler(uint32_t msg_id, uint8_t* data)
-{
-
-    switch (msg_id)
-    {
-    case FRAME0:
-        DriveStateVelocityCanMsgHandler(data);
-        break;
-    case STR_CAN_MSG_ID:
-        DriveStateSteeringCanMsgHandler(data);
-        break;
-
-#ifdef DEBUG
-    case 0x500:
-        StateRequestCanMsgHandler(data);
-        break;
-#endif
-    }
-}
-
-/* CAN TX */
