--- firmware/components/drd/driver/evan_driver.c
@@ -1,3 +1,0 @@
-#include "evan_driver.h"
-
-static int EvanDriverReturn1(void) { return 1; }
