--- firmware/components/drd/driver/evan_driver.h
@@ -1,6 +1,0 @@
-#ifndef __EVAN_APP_H__
-#define __EVAN_APP_H__
-
-#define EVAN_APP 100
-
-#endif //__EVAN_APP_H__
